requirements.txt
# --- Core app ---
streamlit==1.52.2
packaging>=23.2

# --- LangChain ecosystem (pinned for compatibility) ---
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-text-splitters==0.2.4

# --- Vector database ---
chromadb==0.5.3

# --- LLM / Embeddings ---
openai>=1.30.0
tiktoken>=0.7.0

# --- Document loaders ---
pypdf>=4.2.0

# --- Required LangChain runtime dependencies ---
tenacity>=8.2.3
jsonpatch>=1.33

# --- Utilities ---
python-dotenv>=1.0.1
